move `data_input` module to `transaction` module;
export `input` and `data_input` types via in `transaction` module;
