+++ ergo-lib/src/chain/json.rs
@@ -78,9 +78,8 @@ pub mod ergo_box {
 }
 
 pub mod transaction {
-    use crate::chain::transaction::input::Input;
-    use crate::chain::transaction::input::UnsignedInput;
-    use crate::chain::{data_input::DataInput, ergo_box::ErgoBox, transaction::TxId};
+    use crate::chain::transaction::{DataInput, Input, UnsignedInput};
+    use crate::chain::{ergo_box::ErgoBox, transaction::TxId};
     use serde::{Deserialize, Serialize};
 
     #[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
