+++ ergo-lib/src/chain/transaction.rs
@@ -1,14 +1,15 @@
 //! Ergo transaction
 
-pub mod input;
+mod data_input;
+mod input;
 pub mod unsigned;
 
-use self::input::Input;
+pub use data_input::*;
+pub use input::*;
 
 #[cfg(feature = "json")]
 use super::json;
 use super::{
-    data_input::DataInput,
     digest32::{blake2b256_hash, Digest32},
     ergo_box::ErgoBox,
     ergo_box::ErgoBoxCandidate,
