+++ ergo-lib/src/chain/transaction/data_input.rs
@@ -2,7 +2,7 @@
 
 use std::io;
 
-use super::ergo_box::BoxId;
+use crate::chain::ergo_box::BoxId;
 use crate::serialization::{
     sigma_byte_reader::SigmaByteRead, sigma_byte_writer::SigmaByteWrite, SerializationError,
     SigmaSerializable,
